Adition to the delivery system

+++ src/main/java/com/architechz/project/packageservices/models/Package.java
@@ -17,7 +17,7 @@ public class Package {
 
     private double weight;
 
-    private double volume;
+    private BigDecimal volume;
 
     @NotBlank
     private String location;
@@ -36,7 +36,7 @@ public class Package {
     public Package() {
     }
 
-    public Package(Long id, double weight, double volume, @NotBlank String location,
+    public Package(Long id, double weight, BigDecimal volume, @NotBlank String location,
             @NotBlank String type) {
         this.id = id;
         this.weight = weight;
@@ -61,11 +61,11 @@ public class Package {
         this.weight = weight;
     }
 
-    public double getVolume() {
+    public BigDecimal getVolume() {
         return volume;
     }
 
-    public void setVolume(double volume) {
+    public void setVolume(BigDecimal volume) {
         this.volume = volume;
     }
